refactor: split group_column.rs into multiple submoules
